Introduced some typeckecks, moved lazy resolver to its own module
