+++ mypy.ini
@@ -1,0 +1,2 @@
+[mypy]
+disable_error_code = import-untyped, return-value, misc, arg-type, assignment
