+++ pyproject.toml
@@ -14,7 +14,6 @@ classifiers = [
     "Operating System :: OS Independent",
     "Intended Audience :: Developers",
     "Topic :: File Formats",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
@@ -23,7 +22,7 @@ classifiers = [
 dynamic = ["readme", "version"]
 
 dependencies = []
-requires-python = ">= 3.10"
+requires-python = ">= 3.11"
 
 [project.urls]
 Homepage = 'https://github.com/myjniak/hocon'
@@ -50,7 +49,7 @@ packages = [
 content-type = 'text/markdown'
 fragments = [
     {path = "README.md"},
-    {path = "CHANGELOG.md"},
+    {path = "FEATURES.md"},
 ]
 
 [tool.hatch.envs.test]
@@ -65,6 +64,7 @@ dependencies = [
 [tool.hatch.envs.test.scripts]
 coverage = "python -m pytest -vs --cov hocon --cov-fail-under=97 --junitxml=junit/test-results.xml-{env:PYTHON_VERSION}.xml"
 black = "python -m black --check src/"
+mypy = "python -m mypy -p hocon"
 
 [tool.black]
 line-length = 120
