+++ src/hocon/__version__.py
@@ -1,3 +1,3 @@
 __title__ = "hocon"
 __description__ = "A modern HOCON parser with json-like API."
-__version__ = "0.2.0"
+__version__ = "0.3.0"
