+++ src/hocon/parser/_data.py
@@ -1,18 +1,13 @@
 import os
-from dataclasses import dataclass
-from typing import Optional
+from dataclasses import dataclass, field
 
 
 @dataclass
 class ParserInput:
     data: str
     absolute_filepath: str | os.PathLike
-    root_path: Optional[list[str]] = None
+    root_path: list[str] = field(default_factory=list)
     encoding: str = "UTF-8"
-
-    def __post_init__(self):
-        if self.root_path is None:
-            self.root_path = []
 
     def __getitem__(self, item):
         return self.data.__getitem__(item)
